changing Price from double to uint64_t

+++ llob/Types.h
@@ -2,7 +2,7 @@
 
 namespace llob {
 
-using Price = double;
+using Price = uint64_t;
 using InstrumentId = uint16_t;
 using Quantity = uint64_t;
 using OrderId = uint64_t;
